+++ console.go
@@ -16,22 +16,19 @@ type execFunc func(argv0 string, argv []string, envv []string) (err error)
 type exitFunc func(code int)
 
 func console(stdin io.Reader, stdout io.Writer, stderr io.Writer, args []string, exec execFunc, exit exitFunc) {
-	env, program, wait, reportPID := parseConsoleArgs(args)
+	env, program, wait, reportPID := parseConsoleArgs(args, exit)
 
 	if wait {
 		firstByte := make([]byte, 1)
 		readBytes, err := stdin.Read(firstByte)
 		if err != nil {
-			usage(fmt.Sprintf("Failed to read from stdin (%v)", err), false)
-			exit(2)
+			usage(fmt.Sprintf("Failed to read from stdin (%v)", err), true, 2, exit)
 		}
 		if readBytes != 1 {
-			usage("Nothing read from stdin", false)
-			exit(2)
+			usage("Nothing read from stdin", true, 2, exit)
 		}
 		if firstByte[0] != '\000' && firstByte[0] != '\n' {
-			usage("Unexpected first character. Send either a null or a newline character to start program.", false)
-			exit(2)
+			usage("Unexpected first character. Send either a null or a newline character to start program.", true, 2, exit)
 		}
 	}
 
@@ -51,7 +48,7 @@ func console(stdin io.Reader, stdout io.Writer, stderr io.Writer, args []string,
 	}
 }
 
-func parseConsoleArgs(args []string) ([]string, []string, bool, bool) {
+func parseConsoleArgs(args []string, exitFunc func(code int)) ([]string, []string, bool, bool) {
 	var env = os.Environ()
 	var program []string
 	wait := false
@@ -65,20 +62,20 @@ loop:
 		switch args[0] {
 		case "--":
 			if len(args) == 1 {
-				usage("no program to run passed", true)
+				usage("no program to run passed", true, 1, exitFunc)
 
 			}
 			program = args[1:]
 			args = []string{}
 		case "--env":
 			if len(args) == 1 {
-				usage("--env requires an argument", true)
+				usage("--env requires an argument", true, 1, exitFunc)
 			}
 			if strings.HasPrefix(args[1], "--") {
-				usage("--env requires an argument", true)
+				usage("--env requires an argument", true, 1, exitFunc)
 			}
 			if !strings.Contains(args[1], "=") {
-				usage("--env requires an argument in the format of NAME=VALUE", true)
+				usage("--env requires an argument in the format of NAME=VALUE", true, 1, exitFunc)
 			}
 			env = append(env, args[1])
 			args = args[2:]
@@ -89,7 +86,7 @@ loop:
 			reportPID = true
 			args = args[1:]
 		default:
-			usage(fmt.Sprintf("invalid parameter: %s", args[0]), true)
+			usage(fmt.Sprintf("invalid parameter: %s", args[0]), true, 1, exitFunc)
 		}
 	}
 	return env, program, wait, reportPID