+++ wait_signal.go
@@ -7,7 +7,7 @@ import (
 	osSignal "os/signal"
 )
 
-func waitSignal(stderr io.Writer, args []string, exit exitFunc) {
+func waitSignal(stdout io.Writer, args []string, exit exitFunc) {
 	var sig []os.Signal
 	var message = ""
 	var err error
@@ -20,13 +20,13 @@ loop:
 		switch args[0] {
 		case "--message":
 			if len(message) > 0 {
-				usage("--message is duplicate", true)
+				usage("--message is duplicate", true, 1, exit)
 			}
 			message = args[1]
 			args = args[2:]
 		case "--signal":
 			if len(args) < 2 {
-				usage("--signal requires an argument", true)
+				usage("--signal requires an argument", true, 1, exit)
 			}
 			signalName := args[1]
 			var signalObject os.Signal
@@ -34,23 +34,21 @@ loop:
 
 			signalObject, err = processSignalName(signalName)
 			if err != nil {
-				usage(err.Error(), true)
+				usage(err.Error(), true, 1, exit)
 			}
 			sig = append(sig, signalObject)
 		default:
-			usage(fmt.Sprintf("unknown option: %s", args[0]), true)
+			usage(fmt.Sprintf("unknown option: %s", args[0]), true, 1, exit)
 		}
 	}
 	if len(sig) == 0 {
-		usage("missing required --signal option", true)
+		usage("missing required --signal option", true, 1, exit)
 	}
 
 	chanSig := make(chan os.Signal, 1)
 	chanDone := make(chan struct{})
 
-	for _, signalSubscribe := range sig {
-		osSignal.Notify(chanSig, signalSubscribe)
-	}
+	osSignal.Notify(chanSig, sig...)
 
 	go func() {
 		<-chanSig
@@ -59,7 +57,7 @@ loop:
 	<-chanDone
 
 	if len(message) > 0 {
-		fmt.Print(message)
+		_, _ = stdout.Write([]byte(message))
 	}
 	exit(0)
 }